backend/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
pgvector
anthropic
openai
python-dotenv
pydantic-settings
alembic
sentence-transformers
